+++ examples/odata-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/odata-cli
 
+## 0.11.29
+
+### Patch Changes
+
+-   Updated dependencies [24e45780]
+    -   @sap-ux/axios-extension@1.4.1
+    -   @sap-ux/btp-utils@0.11.8
+
 ## 0.11.28
 
 ### Patch Changes