+++ packages/create/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/create
 
+## 0.4.7
+
+### Patch Changes
+
+-   Updated dependencies [24e45780]
+    -   @sap-ux/app-config-writer@0.1.6
+    -   @sap-ux/adp-tooling@0.1.3
+
 ## 0.4.6
 
 ### Patch Changes